requirements.txt
streamlit>=1.35.0
streamlit-authenticator>=0.3.2
streamlit-option-menu>=0.3.12
pyyaml>=6.0
bcrypt>=4.0
requests>=2.31.0
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.22.0
scikit-learn>=1.4.0
lightgbm>=4.3.0
xgboost>=2.0.0
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.29.0
flake8>=7.0.0
